Add another tests

+++ contracts/TokenA.sol
@@ -5,7 +5,7 @@ import "../node_modules/openzeppelin-solidity/contracts/token/ERC20/ERC20.sol";
 
 contract TokenA is Ownable, ERC20 {
 
-  uint256 initSupply = 21 * (10 ** decimals());
+  uint initSupply = 21 * (10 ** decimals());
     
     constructor(
     ) public ERC20("TokenA", "TKNA") {
